Refactor Cooperative Resizer functionality in Window Manager into an extension
